pyproject.toml
[tool.poetry]
name = "kotanima_content"
version = "0.1.0"
description = ""
authors = ["b0und <f1arenova1@gmail.com>"]

[tool.poetry.dependencies]
python = ">=3.9,<3.10"
psycopg2-binary = "^2.8.6"
Pillow = "^8.2.0"
gallery-dl = "^1.17.2"
optimize-images = "^1.4.0"
vk-api = "^11.9.1"
psaw = "^0.1.0"
python-slugify = "^4.0.1"
praw = "^7.2.0"
pytz = "^2021.1"
jikanpy = "^4.3.1"
shikimori-api = "^1.0.1"
pip = "^21.0.1"
install = "^1.3.4"
opencv-python = "^4.5.1"
matplotlib = "^3.4.1"
numba = "^0.53.1"
python-dotenv = "^0.17.0"
yadisk = "^1.2.14"
fire = "^0.4.0"
ImageHash = "^4.2.0"
h5py = "^3.2.1"
pytest-xdist = "^2.3.0"

[tool.poetry.dev-dependencies]
pytest = "^6.2.3"
autopep8 = "^1.5.6"
flake8 = "^3.9.1"
rope = "^0.19.0"
mypy = "^0.812"
black = {version = "^21.4b0", allow-prereleases = true}

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
